Fix the meaning of the `serialrx_inverted` setting for F.Port
`serialrx_inverted` is supposed to be set to OFF for the normal expected
polarity of the signal for the selected protocol. Before this change it
needed to be set to ON when a F.Port receiver was connected directly to
the FC without an inverter or through the inverted S.Port wire present on
some receivers. This change uniformizes the meaning of this setting with
its meaning for the protocols others than F.Port